ECS Update

+++ common/src/ecs/comp_animator.c
@@ -12,7 +12,7 @@ ecs_id ANIMATOR_SYSTEM_SIGNATURE;
 
 comp_animator_constructor_info *ANIMATOR_CONSTRUCTOR_INFO = NULL;
 
-static ecs_ret sys_animator_update(ecs_ecs *ecs, ecs_id *entities, const size_t entity_count, const ecs_dt dt, void *udata) {
+static ecs_ret sys_animator_update(ecs_ecs *ecs, ecs_id *entities, const unsigned int entity_count, const ecs_dt dt, void *udata) {
   if (dt == 0.0f) {
     return 0;
   }
